add CreateRememberedDevice endpoint

+++ models/remembered_device.go
@@ -8,7 +8,7 @@ import (
 type RememberedDevice struct {
 	ID        string `gorm:"primary_key"`
 	OtpID     int    `gorm:"not null"`
-	ExpiresAt int    `gorm:"not null"`
+	ExpiresAt int64  `gorm:"not null"`
 }
 
 func (rd *RememberedDevice) BeforeCreate(db *gorm.DB) error {
